requirements.txt
pyseagull==1.0.0b4
nonogram-solver==0.1

python-dotenv~=1.0.1
numpy~=1.26.4
matplotlib~=3.8.4
python-dateutil~=2.9.0.post0

requests~=2.31.0
beautifulsoup4~=4.12.3